Added Flick gestures

+++ WeatherLite/Weath.cs
@@ -46,12 +46,12 @@ namespace WeatherLite
 
     public class Sys
     {
-        public int type { get; set; }
+        public double type { get; set; }
         public int id { get; set; }
         public double message { get; set; }
         public string country { get; set; }
-        public int sunrise { get; set; }
-        public int sunset { get; set; }
+        public double sunrise { get; set; }
+        public double sunset { get; set; }
     }
 
     public class Weather
@@ -65,8 +65,8 @@ namespace WeatherLite
     public class Main
     {
         public double temp { get; set; }
-        public int pressure { get; set; }
-        public int humidity { get; set; }
+        public double pressure { get; set; }
+        public double humidity { get; set; }
         public double temp_min { get; set; }
         public double temp_max { get; set; }
     }
@@ -74,9 +74,9 @@ namespace WeatherLite
     public class Wind
     {
         public double speed { get; set; }
-        public int deg { get; set; }
-        public int var_beg { get; set; }
-        public int var_end { get; set; }
+        public double deg { get; set; }
+        public double var_beg { get; set; }
+        public double var_end { get; set; }
     }
 
     public class Rain
@@ -86,7 +86,7 @@ namespace WeatherLite
 
     public class Clouds
     {
-        public int all { get; set; }
+        public double all { get; set; }
     }
